diagnosis/requirements.txt
absl-py==0.9.0
asgiref==3.2.7
astor==0.8.1
cycler==0.10.0
dj-database-url==0.5.0
Django==3.0.5
django-heroku==0.3.1
djangorestframework==3.11.0
gast==0.3.3
google-pasta==0.2.0
grpcio==1.28.1
gunicorn==20.0.4
h5py==2.10.0
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.2.1
matplotlib==3.2.1
numpy==1.18.3
opencv-python==4.2.0.34
pandas==1.0.3
Pillow==7.1.1
protobuf==3.11.3
psycopg2==2.8.5
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
sqlparse==0.3.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
Werkzeug==1.0.1
whitenoise==5.0.1
wrapt==1.12.1
django-cors-headers==3.2.1

